% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_transformation.R
\name{add_block_labels}
\alias{add_block_labels}
\title{add_block_labels}
\usage{
add_block_labels(studs, type = c("epoch", "day", "week", "weekday",
  "month", "date"), interval = "start", warning = TRUE)
}
\arguments{
\item{studs}{An \code{SL_tibble} as returned
by the function \code{\link{load_SL_tibble}}.}

\item{type}{A character vector of block label types
to include. Can be one or more of "epoch", "day",
"week", "weekday", "month" and "date". Any block label types that
are not inferrable from the available date-time data are ignored.}

\item{interval}{A character string that decides how block
membership is decided when \code{stude} is of class
\code{interval_SL_tibble}. Can be either "start"
(use \code{start_timestamp}),
"end" (use \code{end_timestamp}) or "middle" (use the midpoint between
\code{start_timestamp} and \code{end_timestamp}).}

\item{warning}{Logical. If \code{TRUE} then a warning is produced
whenever a block label type is not inferrable from the
available date-time data.}
}
\description{
Classify observations from an \code{SL_tibble}
into block labels using available
date-time information. See more information
about "blocks" under the details section.
}
\details{
Block label types can be one or more of "epoch"
(giving labels morning, evening, afternoon and night),
"day" (giving number of days since the first day of the
StudentLife study),
"week" (giving number of weeks since the first week of the
StudentLife study),
"weekday" (giving the day of the week),
"month" (giving number of months since the start of the
StudentLife study) and "date". The StudentLife study
started on 2013-03-24
}
