% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_transformation.R
\name{add_weekdays}
\alias{add_weekdays}
\title{add_weekdays}
\usage{
add_weekdays(studs, day = "day")
}
\arguments{
\item{studs}{A data.frame with a column representing number
of days since the start of the study.}

\item{day}{The name of the column that represents the
number of days since the start of the study.}
}
\description{
Given a data.frame with a column representing days
since the start of the study, creates a column
called "day" that is a factor representing the
day of the week.
}
\examples{
p <- "D:/Datasets/studentlife"
pam <- studentlife::load_SL_tibble(schema = 2, table = 22, location = p)
add_weekdays(timestamp_convert(pam))



}
