% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_transformation.R
\name{add_weeks}
\alias{add_weeks}
\title{add_weeks}
\usage{
add_weeks(studs, timestamp = "resp_time", week1 = 11)
}
\description{
Extract a week variable from the timestamp column
specified by the
\code{timestamp} parameter
}
