% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{csv2tibble}
\alias{csv2tibble}
\title{csv2tibble}
\usage{
csv2tibble(prefix, name, timestamp = "", month1 = 3, week1 = 11,
  day1 = 83, days_in_weeks = FALSE)
}
\arguments{
\item{prefix}{The path prefix.}

\item{name}{The name of the folder in the StudentLife dataset,
a subfolder of the folder specified by \code{prefix}}

\item{timestamp}{Character string indicating which variable
contains unix time stamps to convert to columns date,
day, week and month. Leave as default for no conversion.}

\item{month1}{Integers specifying the
month, week and day of the year that the study begins.
Best left as defaults.}

\item{week1}{Integers specifying the
month, week and day of the year that the study begins.
Best left as defaults.}

\item{day1}{Integers specifying the
month, week and day of the year that the study begins.
Best left as defaults.}

\item{days_in_weeks}{Logical, only used if \code{timestamp}
is not default. If \code{TRUE} then days are per week and
factors rather than per year and integer.}
}
\description{
This function is usually used for importing sensor data from csv
files in the StudentLife dataset. Note that \code{\link{na.exclude}}
is used.
}
\examples{
p <- "C:/Users/danie/Data/StudentLife/dataset/dataset/sensing/"
n <- "conversation"
con1 <- csv2tibble(p,n)
con2 <- csv2tibble(p,n, timestamp = "start_timestamp")
con3 <- csv2tibble(p,n, timestamp = "start_timestamp",
                   days_in_weeks = TRUE)
}
