% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{download_studentlife}
\alias{download_studentlife}
\title{download_studentlife}
\usage{

  download_studentlife(url = "https://studentlife.cs.dartmouth.edu/dataset/dataset.tar.bz2",
  dest = ".", unzip = TRUE, untar = TRUE)
}
\arguments{
\item{url}{The url for the StudentLife dataset.}

\item{dest}{The destination path.}

\item{unzip}{Logical. If \code{TRUE} then the
dataset will be unzipped. Leave as default
unless you plan to do it manually.}

\item{untar}{Logical. If \code{TRUE} then the
dataset will be untarred. Leave as default
unless you plan to do it manually.}
}
\description{
Download the entire StudentLife dataset
}
\examples{
d <- "D:/Datasets/studentlife"
download_studentlife(dest = d)
load_SL_tibble(location = d)

}
