% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_transformation.R
\name{equalise_spacing}
\alias{equalise_spacing}
\title{equalise_spacing}
\usage{
equalise_spacing(studs, FUN = mean, by = c("epoch", "day"),
  per_student = TRUE)
}
\arguments{
\item{studs}{A data.frame or SL_tibble}

\item{FUN}{A function used to aggregate values when multiple
observations are encountered in a time interval.}

\item{by}{A character vector naming one or more
of the columns of \code{studs}. The returned \code{data.frame}
will have one observation (possibly NA) for each unique
value of the columns specified.}

\item{per_student}{A logical. If \code{per_student = FALSE} then
values are aggregated across students.}

\item{y}{An optional character vector specifying one or
more of the columns in \code{studs} to be aggregated by \code{FUN}.}
}
\description{
equalise_spacing
}
