% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_helpers.R
\name{get_EMA_questions}
\alias{get_EMA_questions}
\title{get_EMA_questions}
\usage{
get_EMA_questions(x)
}
\arguments{
\item{x}{A StudentLife tibble whose schema is
EMA, as output by the function
\code{\link[studentlife]{load_SL_tibble}}.}
}
\value{
The EMA_questions attribute of \code{x}
}
\description{
Get the EMA questions from a StudentLife tibble
whose schema is "EMA".
}
