% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_helpers.R
\name{get_schema}
\alias{get_schema}
\title{get_schema}
\usage{
get_schema(x)
}
\arguments{
\item{x}{An object of class StudentLife tibble
(\code{SL_tbl}), as produced by the function
\code{\link[studentlife]{load_SL_tibble}}.}
}
\value{
A character string indicating the schema name
}
\description{
Retrieve the schema name from a StudentLife tibble
}
\examples{
\donttest{
d <- tempdir()
download_studentlife(dest = d)

tab_PAM <- load_SL_tibble(schema = "EMA", table = "PAM", location = d)

# Returns "EMA"
get_schema(tab_PAM)
}

}
