% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_helpers.R
\name{get_table}
\alias{get_table}
\title{get_table}
\usage{
get_table(x)
}
\arguments{
\item{x}{An object of class StudentLife tibble
(\code{SL_tbl}), as produced by the function
\code{\link[studentlife]{load_SL_tibble}}.}
}
\value{
A character string indicating the table name
}
\description{
Retrieve the table name from a StudentLife tibble
}
\examples{
\donttest{
d <- "D:/Datasets/studentlife"
download_studentlife(dest = d)

tab_PAM <- load_SL_tibble(schema = "EMA", table = "PAM", location = d)

# Returns "PAM"
get_table(tab_PAM)
}

}
