% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_transformation.R
\name{group_by_epoch}
\alias{group_by_epoch}
\title{group_by_epoch}
\usage{
group_by_epoch(studs, ..., keep_timestamps = TRUE,
  timestamp = "resp_time")
}
\arguments{
\item{...}{Arguments passed to summarise for group aggregation}

\item{timestamp}{Character string determining what column contains
timestamps for keeping if \code{keep_timestamps = TRUE}}
}
\description{
Group by day and epoch with mean response
}
