% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_helpers.R
\name{is_dateonly_SL_tibble}
\alias{is_dateonly_SL_tibble}
\title{is_dateonly_SL_tibble}
\usage{
is_dateonly_SL_tibble(x)
}
\arguments{
\item{x}{Any object}
}
\value{
Logical
}
\description{
Confirm that an object is a
date-only StudentLife tibble
}
\examples{
d <- tempdir()
download_studentlife(location = d, url = "testdata")

tab_DL <- load_SL_tibble(
  schema = "education", table = "deadlines", location = d)

# Returns TRUE
is_dateonly_SL_tibble(tab_DL)

}
