% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_helpers.R
\name{is_interval_SL_tbl}
\alias{is_interval_SL_tbl}
\title{is_interval_SL_tbl}
\usage{
is_interval_SL_tbl(x)
}
\arguments{
\item{x}{Any object}
}
\value{
Logical
}
\description{
Confirm that an object is
an interval StudentLife tibble
}
\examples{
\donttest{
d <- "D:/Datasets/studentlife"
download_studentlife(dest = d)

tab_con <- load_SL_tibble(
  schema = "sensing", table = "conversation", location = d, csv_nrow = 10)

# Returns TRUE
is_interval_SL_tbl(tab_con)
}

}
