% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_helpers.R
\name{is_interval_SL_tibble}
\alias{is_interval_SL_tibble}
\title{is_interval_SL_tibble}
\usage{
is_interval_SL_tibble(x)
}
\arguments{
\item{x}{Any object}
}
\value{
Logical
}
\description{
Confirm that an object is
an interval StudentLife tibble
}
\examples{
d <- tempdir()
download_studentlife(location = d, url = "testdata")

tab_con <- load_SL_tibble(
  schema = "sensing", table = "conversation", location = d, csv_nrow = 10)

# Returns TRUE
is_interval_SL_tibble(tab_con)

}
