% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_helpers.R
\name{is_timestamp_SL_tibble}
\alias{is_timestamp_SL_tibble}
\title{is_timestamp_SL_tibble}
\usage{
is_timestamp_SL_tibble(x)
}
\arguments{
\item{x}{Any object}
}
\value{
Logical
}
\description{
Confirm that an object is a
timestamped StudentLife tibble
}
\examples{
d <- tempdir()
download_studentlife(location = d, url = "testdata")

tab_PAM <- load_SL_tibble(schema = "EMA", table = "PAM", location = d)

# Returns TRUE
is_timestamp_SL_tibble(tab_PAM)

}
