% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{load_SL_tibble}
\alias{load_SL_tibble}
\title{load_SL_tibble}
\usage{
load_SL_tibble(schema, table, location = ".",
  time_options = c("period", "timestamp", "dateonly", "dateless"), vars,
  csv_nrows, datafolder = "/dataset")
}
\arguments{
\item{schema}{An integer. The menu 1 choice. Leave
blank to choose interactively.}

\item{table}{An integer. The menu 2 choice. Leave
blank to choose interactively.}

\item{location}{The path to a copy of the StudentLife dataset.}

\item{time_options}{A character vector specifying which
table types to include in the menu, organised by level
of time information present. The default includes everything. Note
this parameter only has an effect when used with the interactive menu.}

\item{vars}{Character vector of variable
names to import for all students. Leave
blank and this will be chosen interactively
if necesssary.}

\item{csv_nrows}{An integer specifying the number of rows to read
per student if the target is a csv. The largest files in StudentLife are csv
files, so this allows code testing with less overhead.}

\item{datafolder}{Specifies the subfolder of \code{location}
that contains the relevant data. This should normally
be left as the default.}
}
\description{
Import a chosen StudentLife table as
a tibble. Leave \code{schema} and \code{table}
unspecified to choose interactively via a
menu.
}
\examples{
p <- "D:/Datasets/studentlife"
students <- load_SL_tibble(location = p)

}
