% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_transformation.R
\name{make_days_in_weeks}
\alias{make_days_in_weeks}
\title{make_days_in_weeks}
\usage{
make_days_in_weeks(studs, day = "day")
}
\arguments{
\item{studs}{A data.frame with a column representing number
of days since the start of the study.}

\item{day}{The name of the column that represents the
number of days since the start of the study.}
}
\description{
Given a data.frame with a column representing days
since the start of the study, creates a column
called "day" that is a factor representing the
day of the week.
}
\examples{
p <- "C:/Users/danie/Data/StudentLife/dataset/dataset"
pam <- studentlife::read_from_SL(menu1 = 2, menu2 = 22, prefix = p)
make_days_in_weeks(timestamp_convert(pam))



}
