% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_transformation.R
\name{PAM_categorise}
\alias{PAM_categorise}
\title{PAM_categorise}
\usage{
PAM_categorise(studs, pam_name = "picture_idx", types = c("quadrant",
  "valence", "arousal"))
}
\arguments{
\item{studs}{A data.frame with a column representing
Photographic Affect Meter (PAM) score.}

\item{pam_name}{Character. The name of the column
representing PAM.}

\item{types}{Character vector containing the categories,
one or more of "quadrant", "valence" and "arousal" into
which to code PAM scores.}
}
\description{
Categorise Photographic Affect Meter (PAM) scores into
4 categories by either PAM Quadrant, Valence or Arousal
(or multiple of these).
}
\details{
The 4 Quadrant categories are as follows:
Quadrant 1: low valence, low arousal.
Quadrant 2: low valence, high arousal.
Quadrant 3: high valence, low arousal.
Quadrant 4: high valence, high arousal.

The 4 valence / arousal categories are: low (1),
med-low (2), high-low (3), high (4).
}
