% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_from_SL}
\alias{read_from_SL}
\title{read_from_SL}
\usage{
read_from_SL(menu1, menu2, prefix = ".", vars)
}
\arguments{
\item{menu1}{The Menu 1 choice. Leave
blank to choose interactively.}

\item{menu2}{The Menu 1 choice. Leave
blank to choose interactively.}

\item{prefix}{The path to the top
directory of the StudentLife dataset}

\item{vars}{Character vector of variable
names to import for all students. Leave
blank and this will be chosen interactively
if necesssary.}
}
\description{
Import a chosen StudentLife data set as
a tibble. Leave \code{menu1} and \code{menu2}
unspecified to choose interactively via a
menu.
}
\examples{
pr <- "C:/Users/danie/Data/StudentLife/dataset/dataset"
students <- import_from_SL(prefix = prefix)

}
