% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_tbl_explore.R
\name{response_hour_hist}
\alias{response_hour_hist}
\title{response_hour_hist}
\usage{
response_hour_hist(tab, break_hours = 10, xlab = "Hours into study",
  main = paste0("Distribution of ", attr(tab, "table"),
  " response times"), ...)
}
\arguments{
\item{tab}{A timestamp_SL_tbl or interval_SL_tbl.}

\item{break_hours}{Specify the width in hours of each
histogram bin.}

\item{xlab}{Argument passed to \code{\link[graphics]{hist}}.}

\item{main}{Argument passed to \code{\link[graphics]{hist}}.}

\item{...}{Arguments passed to \code{\link[graphics]{hist}}.}
}
\description{
Produce a histogram with frequency of observations
in each hour or number of hours
}
