% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SL_tables}
\alias{SL_tables}
\title{List of all the tables available in the StudentLife dataset.}
\format{An object of class \code{list} of length 5.}
\source{
\url{https://studentlife.cs.dartmouth.edu/}
}
\usage{
SL_tables
}
\description{
The list elements are character vectors of table names
and the element names correspond to schema names.
}
\keyword{datasets}
