% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{strongest_uids}
\alias{strongest_uids}
\title{strongest_uids}
\usage{
strongest_uids(studs, n = 10)
}
\description{
Returns the uids of the n students in the
data.frame that have the most responses
}
