% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_transformation.R
\name{timestamp_convert}
\alias{timestamp_convert}
\title{timestamp_convert}
\usage{
timestamp_convert(studs, month1 = 3, week1 = 11, day1 = 83,
  days_in_weeks = FALSE, timestamp = "resp_time")
}
\arguments{
\item{studs}{A data.frame where each row corresponds to an
observation on a student, and at least one column contains
UNIX timestamps.}

\item{month1, week1, day1}{Integers specifying the
month, week and day of the year that the study begins.
Best left as defaults.}

\item{days_in_weeks}{Logical, only used if \code{timestamp}
is not default. If \code{TRUE} then days are per week and
factors rather than per year and integer.}

\item{timestamp}{Character string indicating which variable
contains unix time stamps to convert to columns date,
day, week and month. Leave as default for no conversion.}
}
\description{
Convert StudentLife UNIX timestamps to date, day, week and month
in study.
}
\examples{
p <- "C:/Users/danie/Data/StudentLife/dataset/dataset"
pam <- studentlife::read_from_SL(menu1 = 2, menu2 = 22, prefix = p)

timestamp_convert(pam)
# Compare to
timestamp_convert(pam, days_in_weeks = TRUE)

}
