% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_tbl_explore.R
\name{vis_NAs}
\alias{vis_NAs}
\title{vis_NAs}
\usage{
vis_NAs(tab, response, main = paste0("Missing values by student (",
  attr(tab, "table"), ")"))
}
\arguments{
\item{tab}{A an object of class \code{reg_SL_tbl} as
produced by the function \code{\link[studentlife]{regularise_time}}.}

\item{response}{A character string naming one of the columns
in \code{tab} that is not in \code{attr(tab, "blocks")}. If
missing then this defaults to the first such column name.}

\item{main}{The plot title, passed to \code{\link[ggplot2]{ggtitle}}.}
}
\value{
A ggplot object.
}
\description{
Produce a visualisation of the number of missing values
among each student in a regularised SL_tbl.
}
\examples{
\donttest{
d <- "D:/Datasets/studentlife"
download_studentlife(dest = d)

tab_PAM <- load_SL_tibble(schema = "EMA", table = "PAM", location = d)

vis_NAs(reg_PAM, response = "m")
}

}
