% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_tbl_explore.R
\name{vis_NAs_by_student}
\alias{vis_NAs_by_student}
\title{vis_NAs_by_student}
\usage{
vis_NAs_by_student(tab)
}
\description{
This function should be applied after adding
NAs to the imported data with \code{\link{add_NAs}}
}
