% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{vis_stud_NAs}
\alias{vis_stud_NAs}
\title{vis_stud_NAs}
\usage{
vis_stud_NAs(studs)
}
\description{
This function should be applied after adding
NAs to the imported data with \code{\link{add_NAs}}
}
