% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{csv2tibble}
\alias{csv2tibble}
\title{csv2tibble}
\usage{
csv2tibble(prefix, name, timestamp, month1 = 3, week1 = 11,
  day1 = 83)
}
\arguments{
\item{prefix}{The path prefix}

\item{name}{The name of the folder in the StudentLife dataset,
a subfolder of the folder specified by \code{prefix}}

\item{timestamp}{Character string indicating which variable
contains unix time stamps to convert to columns date,
day, week and month. Leave unspecified for no conversion.}
}
\description{
This function is usually used for importing sensor data from csv
files in the StudentLife dataset. Note that \code{\link{na.exclude}}
is used.
}
\examples{
p <- "C:/Users/danie/Data/StudentLife/dataset/dataset/sensing/"
n <- "conversation"
csv2tibble(p,n)
csv2tibble(p,n, timestamp = "start_timestamp")

}
