% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{JSON2tibble}
\alias{JSON2tibble}
\title{JSON2tibble}
\usage{
JSON2tibble(prefix, name, vars, timestamp, month1 = 3, week1 = 11,
  day1 = 83)
}
\arguments{
\item{prefix}{The path prefix}

\item{name}{The name of the folder in the StudentLife dataset,
usually as a subfolder of the EMA/response directory}

\item{vars}{Vector of names of variables to read for each student}

\item{timestamp}{Character string indicating which variable
contains unix time stamps to convert to columns date,
day, week and month. Leave unspecified for no conversion.}

\item{month1, week1, day1}{Integers specifying the
month, week and day of the year that the study begins.
Best left as defaults.}
}
\description{
This function is mostly used for reading StudentLife EMA data into R.
Note that \code{\link{na.exclude}} is used. Only the columns specified
by \code{var} will be selected.
}
\examples{
# Import Stress EMAs
p <- "C:/Users/danie/Data/StudentLife/dataset/dataset/EMA/response/"
n <-  "Stress"
v <- c("level", "resp_time")
stress <- JSON2tibble(p, n, v)

# Import PAM EMAs
p <- "C:/Users/danie/Data/StudentLife/dataset/dataset/EMA/response/"
n <-  "PAM"
v <- c("picture_idx", "resp_time")
pam <- JSON2tibble(p, n, v)
pam <- JSON2tibble(p, n, v, timestamp = "resp_time")

}
