% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SL_tables}
\alias{SL_tables}
\title{List of all the tables available in the StudentLife dataset.}
\format{
An object of class \code{list} of length 5.
}
\source{
\url{https://studentlife.cs.dartmouth.edu/}
}
\usage{
SL_tables
}
\description{
This command returns a 5 element list. Each of the five elements
are given names corresponding to the schema names of the studentlife
data set. Each element is a vector of strings, where each string
corresponds to the name of a table within the respective schema.
}
\keyword{datasets}
