% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentlife.R
\docType{package}
\name{studentlife}
\alias{studentlife}
\alias{studentlife-package}
\title{Investigating the StudentLife dataset from Dartmouth}
\description{
Seamless tidy handling and convenient navigation
of a rich mHealth dataset
}
\details{
Details on the StudentLife dataset are available at
https://studentlife.cs.dartmouth.edu/.
}
\section{Update}{
 Current updates are available through
URL: https://github.com/frycast/studentlife
}

\section{BugReports}{
 https://github.com/frycast/studentlife/issues
}

\author{
Daniel Fryer \email{d.fryer@latrobe.edu.au}
}
